+++ src/main/java/com/librarymanager/repository/SocioRepository.java
@@ -1,11 +1,11 @@
 package com.librarymanager.repository;
 
+import com.librarymanager.model.Socio;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.librarymanager.model.Socio;
-
 @Repository
-public interface SocioRepository extends JpaRepository<Socio, Integer> {
+public interface SocioRepository extends JpaRepository<Socio, Long> {
+    // Se precisar de consultas customizadas, pode adicionar aqui.
 }
